refactor(ui): extract slide_list_utils to keep slide_list.rs under 1000 lines
Move format_multiline, field_has_warning, slide_has_any_warning,
and apply_focused_class to slide_list_utils.rs module.
